+++ source/funcs.py
@@ -1,0 +1,295 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Jun 11 21:21:17 2018
+
+@author: Lully
+Fonctions et classes génériques pour Bibliostratus
+"""
+
+from unidecode import unidecode
+
+#Quelques listes de signes à nettoyer
+listeChiffres = ["0","1","2","3","4","5","6","7","8","9"]
+lettres = ["a","b","d","e","f","g","h","i","j","k","l","m","n","o","p","q","r","s","t","u","v","w","x","y","z"]
+lettres_sauf_x = ["a","b","d","e","f","g","h","i","j","k","l","m","n","o","p","q","r","s","t","u","v","w","y","z"]
+ponctuation = [".",",",";",":","?","!","%","$","£","€","#","\\","\"","&","~","{","(","[","`","\\","_","@",")","]","}","=","+","*","\/","<",">",")","}"]
+
+def nettoyage(string,remplacerEspaces=True,remplacerTirets=True):
+    """nettoyage des chaines de caractères (titres, auteurs, isbn) : suppression ponctuation, espaces (pour les titres et ISBN) et diacritiques"""
+    string = unidecode(string.lower())
+    for signe in ponctuation:
+        string = string.replace(signe,"")
+    string = string.replace("'"," ")
+    if (remplacerTirets == True):
+        string = string.replace("-"," ")
+    if (remplacerEspaces == True):
+        string = string.replace(" ","")
+    return string
+
+
+def nettoyage_lettresISBN(isbn):
+    isbn = unidecode(isbn.lower())
+    char_cle = "0123456789xX"
+    for signe in ponctuation:
+        isbn = isbn.replace(signe,"")
+    prefix = isbn[0:-1]
+    cle = isbn[-1]
+    for lettre in lettres:
+        prefix = prefix.replace(lettre, "")
+    if (cle in char_cle):
+        cle = cle.upper()
+    else:
+        cle = ""
+    return prefix+cle
+
+def nettoyageIsbnPourControle(isbn):
+    isbn = nettoyage(isbn)
+    if (isbn != ""):
+        isbn = nettoyage_lettresISBN(isbn)
+    if (len(isbn) < 10):
+        isbn = ""
+    elif (isbn[0:3] == "978" or isbn[0:3] == "979"):
+        isbn = isbn[3:12]
+    else:
+        isbn = isbn[0:10]
+    return isbn
+
+def nettoyageIssnPourControle(issn):
+    issn = nettoyage(issn).replace(" ","")
+    if (issn != ""):
+        issn = nettoyage_lettresISBN(issn)
+    if (len(issn) < 8):
+        issn = ""
+    else:
+        issn = issn[0:8]
+    return issn
+
+def nettoyageAuteur(auteur,justeunmot=True):
+    listeMots = [" par "," avec "," by "," Mr. "," M. "," Mme "," Mrs "]
+    for mot in listeMots:
+        auteur = auteur.replace(mot,"")
+    for chiffre in listeChiffres:
+        auteur = auteur.replace(chiffre,"")
+    auteur = nettoyage(auteur.lower(),False)
+    auteur = auteur.split(" ")
+    auteur = sorted(auteur,key=len,reverse=True)
+    auteur = [auteur1 for auteur1 in auteur if len(auteur1) > 1]
+    if (auteur is not None and auteur != []):
+        if (justeunmot==True):
+            auteur = auteur[0]
+        else:
+            auteur = " ".join(auteur)
+    else:
+        auteur = ""
+    return auteur
+
+def nettoyageTitrePourControle(titre):
+    titre = nettoyage(titre,True)
+    return titre
+    
+
+def nettoyageDate(date):
+    date = unidecode(date.lower())
+    for lettre in lettres:
+        date = date.replace(lettre,"")
+    for signe in ponctuation:
+        date = date.split(signe)
+        date = " ".join(annee for annee in date if annee != "")
+    return date
+
+def nettoyageTome(numeroTome):
+    if (numeroTome):
+        numeroTome = unidecode(numeroTome.lower())
+        for lettre in lettres:
+            numeroTome = numeroTome.replace(lettre,"")
+        for signe in ponctuation:
+            numeroTome = numeroTome.split(signe)
+            numeroTome = "~".join(numero for numero in numeroTome)
+        numeroTome = numeroTome.split("~")
+        numeroTome = [numero for numero in numeroTome if numero != ""]
+        if (numeroTome != []):
+            numeroTome = numeroTome[-1]
+        numeroTome = ltrim(numeroTome)
+    return numeroTome
+    
+def nettoyagePubPlace(pubPlace) :
+    """Nettoyage du lieu de publication"""
+    pubPlace = unidecode(pubPlace.lower())
+    for chiffre in listeChiffres:
+        pubPlace = pubPlace.replace(chiffre,"")
+    for signe in ponctuation:
+        pubPlace = pubPlace.split(signe)
+        pubPlace = " ".join(mot for mot in pubPlace if mot != "")
+    return pubPlace
+
+def ltrim(nombre_texte):
+    "Supprime les 0 initiaux d'un nombre géré sous forme de chaîne de caractères"
+    while(len(nombre_texte) > 1 and nombre_texte[0] == "0"):
+        nombre_texte = nombre_texte[1:]
+    return nombre_texte
+
+def nettoyage_isbn(isbn):
+    isbn_nett = isbn.split(";")[0].split(",")[0].split("(")[0].split("[")[0]
+    isbn_nett = isbn_nett.replace("-","").replace(" ","")
+    isbn_nett = unidecode(isbn_nett)
+    for signe in ponctuation:
+        isbn_nett = isbn_nett.replace(signe,"")
+    isbn_nett = isbn_nett.lower()
+    for lettre in lettres_sauf_x:
+        isbn_nett = isbn_nett.replace(lettre,"")
+    return isbn_nett
+
+def conversionIsbn(isbn):
+    longueur = len(isbn)
+    isbnConverti = ""
+    if (longueur == 10):
+        isbnConverti = conversionIsbn1013(isbn)
+    elif (longueur == 13):
+        isbnConverti = conversionIsbn1310(isbn)
+    return isbnConverti
+
+#conversion isbn13 en isbn10
+def conversionIsbn1310(isbn):
+    if (isbn[0:3] == "978"):
+        prefix = isbn[3:-1]
+        check = check_digit_10(prefix)
+        return prefix + check
+    else:
+        return ""
+
+#conversion isbn10 en isbn13
+def conversionIsbn1013(isbn):
+    prefix = '978' + isbn[:-1]
+    check = check_digit_13(prefix)
+    return prefix + check
+    
+def check_digit_10(isbn):
+    assert len(isbn) == 9
+    sum = 0
+    for i in range(len(isbn)):
+        c = int(isbn[i])
+        w = i + 1
+        sum += w * c
+    r = sum % 11
+    if (r == 10):
+        return 'X'
+    else: 
+        return str(r)
+
+def check_digit_13(isbn):
+    assert len(isbn) == 12
+    sum = 0
+    for i in range(len(isbn)):
+        c = int(isbn[i])
+        if (i % 2):
+            w = 3
+        else: 
+            w = 1
+        sum += w * c
+    r = 10 - (sum % 10)
+    if (r == 10):
+        return '0'
+    else:
+        return str(r)
+
+
+numeral_map = tuple(zip(
+    (1000, 900, 500, 400, 100, 90, 50, 40, 10, 9, 5, 4, 1),
+    ('M', 'CM', 'D', 'CD', 'C', 'XC', 'L', 'XL', 'X', 'IX', 'V', 'IV', 'I')
+))
+
+def int_to_roman(i):
+    """Conversion de chiffres arabers en chiffres romains"""
+    result = []
+    for integer, numeral in numeral_map:
+        count = i // integer
+        result.append(numeral * count)
+        i -= integer * count
+    return ''.join(result)
+
+def roman_to_int(n):
+    """Consersion de chiffres romains en chiffres arabes"""
+    i = result = 0
+    for integer, numeral in numeral_map:
+        while n[i:i + len(numeral)] == numeral:
+            result += integer
+            i += len(numeral)
+    return result
+
+def convert_volumes_to_int(n):
+    for char in ponctuation:
+        n = n.replace(char,"-")
+    n = n.replace(" ","-")
+    liste_n = [e for e in n.split("-") if e != ""]
+    liste_n_convert = []
+    for n in liste_n:
+        try:
+            int(n)
+            liste_n_convert.append(n)
+        except ValueError:
+            c = roman_to_int(n)
+            if (c != 0):
+                liste_n_convert.append(c)
+    liste_n_convert = set(ltrim(str(nb)) for nb in liste_n_convert if nb != "")
+    n_convert = " ".join([str(el) for el in list(liste_n_convert)])
+    return n_convert
+
+
+
+class Bib_record: 
+    """Classe définissant les propriétés d'une notice mise en entrée 
+    du module d'alignemen
+    - son nom
+    - son prénom
+    - son âge
+    - son lieu de résidence"""
+
+    def __init__(self, input_row, option_record): # Notre méthode constructeur
+        """Pour l'instant, on ne va définir qu'un seul attribut"""
+        self.NumNot = input_row[0]
+        self.frbnf = input_row[1]
+        self.ark_init = input_row[2]
+        self.isbn_init = ""
+        self.ean = ""
+        self.titre = ""
+        self.auteur = ""
+        self.date = ""
+        self.tome = ""
+        self.publisher = ""
+        self.comref = ""
+        self.issn = ""
+        self.pubPlace = ""
+        if (option_record == 1):
+            self.isbn_init = input_row[3]
+            self.ean = input_row[4]
+            self.titre = input_row[5]
+            self.auteur = input_row[6]
+            self.date = input_row[7]
+            self.tome = input_row[8]
+            self.publisher = input_row[9]
+        if (option_record == 2 or option_record == 3):
+            self.ean = input_row[3]
+            self.comref = input_row[4]
+            self.titre = input_row[5]
+            self.auteur = input_row[6]
+            self.date = input_row[7]
+            self.publisher = input_row[8]
+        if (option_record == 4):
+            self.issn = input_row[3]
+            self.titre = input_row[4]
+            self.auteur = input_row[5]
+            self.date = input_row[6]
+            self.pubPlace = input_row[7]
+        self.isbn_nett = nettoyageIsbnPourControle(self.isbn_init)
+        self.isbn_propre = nettoyage_isbn(self.isbn_init)
+        self.isbn_converti = conversionIsbn(self.isbn_propre)
+        self.ean_nett = nettoyageIsbnPourControle(self.ean)
+        self.ean_propre = nettoyage_isbn(self.ean)
+        self.titre_nett= nettoyageTitrePourControle(self.titre)
+        self.auteur_nett = nettoyageAuteur(self.auteur, False)
+        self.date_nett = nettoyageDate(self.date)
+        self.tome_nett = convert_volumes_to_int(self.tome)
+        self.publisher_nett = nettoyageAuteur(self.publisher, False)
+        if (self.publisher_nett == ""):
+            self.publisher_nett = self.publisher
+        self.pubPlace_nett = nettoyagePubPlace(self.pubPlace)